refactor : config 설정
