+++ src/main/java/com/crude/travelcrew/global/config/JpaConfig.java
@@ -2,8 +2,10 @@ package com.crude.travelcrew.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @Configuration
 @EnableJpaAuditing
+@EnableJpaRepositories(basePackages = {"com.crude.travelcrew.domain.*.repository"})
 public class JpaConfig {
 }
